src/python/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
moviepy>=1.0.3